requirements.txt
numpy>=1.24
scipy>=1.11
sounddevice>=0.4.6
librosa>=0.10
matplotlib>=3.8
pandas>=2.1
click>=8.1
pyyaml>=6.0
